Adjust planetator to new imgui-rs

+++ planetator/src/savegames.rs
@@ -6,7 +6,7 @@ pub enum Savegame {
     Version0 {
         generator: String,
         colorator: String,
-        select_channels: Vec<(String, i32)>,
+        select_channels: Vec<(String, usize)>,
         active_textures: Vec<(String, (i32, i32), Vec<u8>)>,
     }
 }
